refactor: deduplicate shared driver and TUI helpers
Move supervisor_image_tag/supervisor_image_should_refresh from
openshell-driver-docker and openshell-driver-podman into
openshell-core::driver_utils, which already hosts shared driver
utilities.  Both drivers now import the single canonical copy.

Replace the private unix_epoch_millis() function in
openshell-server::ssh_sessions with openshell_core::time::now_ms,
which exists precisely to avoid local reimplementations of this
pattern.

Extract a draw_empty_message() helper into openshell-tui::ui::mod
and replace the six copy-pasted Rect+Paragraph blocks across
dashboard, sandboxes, providers (x2), global_settings, and
sandbox_settings with calls to the shared helper.
