+++ crates/openshell-core/src/driver_utils.rs
@@ -89,3 +89,38 @@ pub fn sandbox_log_level(sandbox: &DriverSandbox, default_level: &str) -> String
         .unwrap_or(default_level)
         .to_string()
 }
+
+// ---------------------------------------------------------------------------
+// Supervisor image helpers (shared by Docker and Podman drivers)
+// ---------------------------------------------------------------------------
+
+/// Return the tag portion of a supervisor image reference, or `None` if the
+/// reference uses a digest (`@sha256:...`).
+///
+/// Examples:
+/// - `"ghcr.io/org/image:1.2.3"` → `Some("1.2.3")`
+/// - `"ghcr.io/org/image:latest"` → `Some("latest")`
+/// - `"ghcr.io/org/image"` → `Some("latest")`  (implied tag)
+/// - `"ghcr.io/org/image@sha256:abc"` → `None`  (pinned by digest)
+/// - `"ghcr.io/org/image:"` → `None`  (empty tag)
+pub fn supervisor_image_tag(image: &str) -> Option<&str> {
+    if image.contains('@') {
+        return None;
+    }
+
+    let image_name = image.rsplit('/').next().unwrap_or(image);
+    image_name
+        .rsplit_once(':')
+        .map_or(Some("latest"), |(_, tag)| {
+            if tag.is_empty() { None } else { Some(tag) }
+        })
+}
+
+/// Return `true` if the supervisor image should be refreshed before each use.
+///
+/// Mutable tags (`dev`, `latest`) are always re-pulled so that the running
+/// container tracks the latest pushed version.  Digest-pinned references and
+/// all other versioned tags are treated as immutable and pulled at most once.
+pub fn supervisor_image_should_refresh(image: &str) -> bool {
+    matches!(supervisor_image_tag(image), Some("dev" | "latest"))
+}
