+++ crates/openshell-server/src/ssh_sessions.rs
@@ -5,6 +5,7 @@
 
 use openshell_core::ObjectId;
 use openshell_core::proto::SshSession;
+use openshell_core::time::now_ms;
 use prost::Message;
 use std::sync::Arc;
 use std::time::Duration;
@@ -33,7 +34,7 @@ pub fn spawn_session_reaper(store: Arc<Store>, interval: Duration) {
 }
 
 async fn reap_expired_sessions(store: &Store) -> Result<(), String> {
-    let now_ms = unix_epoch_millis();
+    let now_ms = now_ms();
 
     let records = store
         .list(SshSession::object_type(), 1000, 0)
@@ -68,16 +69,6 @@ async fn reap_expired_sessions(store: &Store) -> Result<(), String> {
     Ok(())
 }
 
-fn unix_epoch_millis() -> i64 {
-    i64::try_from(
-        std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_millis(),
-    )
-    .unwrap_or(i64::MAX)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -101,10 +92,6 @@ mod tests {
             expires_at_ms,
             revoked,
         }
-    }
-
-    fn now_ms() -> i64 {
-        unix_epoch_millis()
     }
 
     #[tokio::test]
